requirements.txt
Crawl4AI==0.4.247
python-dotenv==1.0.1
pydantic==2.10.6
fastapi==0.104.1
uvicorn[standard]==0.24.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
boto3==1.34.0
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
playwright
